requirements.txt
PyYAML~=6.0.1
pandas~=2.1.4
ping3~=4.0.8
tqdm~=4.67.1